Updated migrations with relations.

+++ database/migrations/2017_05_17_065034_create_students_table.php
@@ -17,7 +17,7 @@ class CreateStudentsTable extends Migration
             $table->increments('id');
             $table->string('firstname');
             $table->string('lastname');
-            $table->mediumInteger('studentnumber');
+            $table->integer('studentnumber');
             $table->timestamps();
         });
     }
